% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_api_spec.R
\name{api_spec}
\alias{api_spec}
\title{Update the OpenAPI specification from model metadata}
\usage{
api_spec(spec, args)
}
\arguments{
\item{spec}{An OpenAPI Specification formatted list object}

\item{args}{A list of model metadata including \code{other_pinned}, \code{meta},
\code{ptype}, and \code{path}}
}
\value{
The new OpenAPI Specification object updated
}
\description{
Update the OpenAPI specification from model metadata
}
