% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{augment.vetiver_endpoint}
\alias{augment.vetiver_endpoint}
\title{Post new data to a deployed model API endpoint and augment with predictions}
\usage{
\method{augment}{vetiver_endpoint}(x, new_data, ...)
}
\arguments{
\item{x}{A model API endpoint object created with \code{\link[=vetiver_endpoint]{vetiver_endpoint()}}.}

\item{new_data}{New data for making predictions, such as a data frame.}

\item{...}{Extra arguments passed to \code{\link[httr:POST]{httr::POST()}}}
}
\value{
The \code{new_data} with added prediction column(s).
}
\description{
Post new data to a deployed model API endpoint and augment with predictions
}
\examples{

if (FALSE) {
endpoint <- vetiver_endpoint("http://127.0.0.1:8088/predict")
augment(endpoint, mtcars[4:7, -1])
}

}
\seealso{
\code{\link[=predict.vetiver_endpoint]{predict.vetiver_endpoint()}}
}
