% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer.R
\name{get_term_levels}
\alias{get_term_levels}
\title{Get a reference grid-like object from all model terms}
\usage{
get_term_levels(terms, data)
}
\arguments{
\item{terms}{character vector}

\item{data}{data.frame}
}
\description{
Get a reference grid-like object from all model terms
}
