% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-docker.R
\name{glue_sys_reqs}
\alias{glue_sys_reqs}
\title{Generate the Docker command required to install system dependencies for given packages}
\usage{
glue_sys_reqs(pkgs, distribution = "ubuntu", release = "20.04")
}
\arguments{
\item{pkgs}{A character vector of packages for which system dependencies are
to be determined.}

\item{distribution}{Linux distribution for which the packages will be
installed. This affects both the system dependencies, and the installation
command. Currently only "ubuntu" and "centos" are supported.}

\item{release}{The release of the Linux distribution, eg. "20.04" for
Ubuntu Focal Fossa or "7" for CentOS 7.}
}
\value{
A character string which comprises a "RUN" command for a Dockerfile
}
\description{
Generate the Docker command required to install system dependencies for given packages
}
\keyword{internal}
