% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelops-pr-deploy.R
\name{handle_model}
\alias{handle_model}
\alias{handle_model.default}
\alias{handle_model.lm}
\title{Wrapper function for creating model handler functions}
\usage{
handle_model(x, ...)

\method{handle_model}{default}(x, ...)

\method{handle_model}{lm}(x, other, type)
}
\arguments{
\item{x}{A trained model stored in a \code{\link[=modelops]{modelops()}} object}

\item{...}{Other arguments, not currently used.}

\item{other}{Other components of a \code{\link[=modelops]{modelops()}} object}

\item{type}{A single character or \code{NULL} describing the type of prediction.
The specific pinned model must support the \code{type} requested. Some examples
of \code{type} for \code{\link[workflows:workflow]{workflows::workflow()}} include "class", "prob", and "numeric".
When \code{NULL}, a default prediction type will be chosen based on the model
characteristics.}
}
\description{
The \code{handle_model()} function for each model type returns a list of
functions:
\itemize{
\item The \code{on_start} function executes when the API starts. Use this function
for tasks like loading packages.
\item The \code{handler} function executes at each API call. Use this function for
calling \code{predict()} and other tasks that must be executed at each API call.
}
}
