% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme.R, R/lmer.R
\name{handler_predict.lme}
\alias{handler_predict.lme}
\title{Title}
\usage{
\method{handler_predict}{lme}(vetiver_model, ...)

\method{handler_predict}{lme}(vetiver_model, ...)
}
\arguments{
\item{...}{}
}
\description{
Title

Title
}
