% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{model_endpoint}
\alias{model_endpoint}
\title{Create a model API endpoint object for prediction}
\usage{
model_endpoint(url, ...)
}
\arguments{
\item{url}{A trained model}

\item{...}{Other arguments, not currently used.}
}
\value{
A new \code{model_endpoint} object
}
\description{
Create a model API endpoint object for prediction
}
\examples{

\dontrun{
model_endpoint("https://colorado.rstudio.com/rsc/traffic-crashes/predict")
}

}
