% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin-model.R
\name{model_pinner}
\alias{model_pinner}
\alias{model_pinner.default}
\alias{model_pinner.lm}
\title{Wrapper function for pinning a model to a board of models}
\usage{
model_pinner(x, ...)

\method{model_pinner}{default}(x, ...)

\method{model_pinner}{lm}(x, ...)
}
\arguments{
\item{x}{A trained model}

\item{...}{Other arguments passed from \code{\link[=pin_model]{pin_model()}}}
}
\description{
Wrapper function for pinning a model to a board of models
}
