% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelops_dashboard.R
\name{modelops_dashboard}
\alias{modelops_dashboard}
\title{R Markdown format for model monitoring dashboards}
\usage{
modelops_dashboard(..., display_pins = TRUE, display_api = TRUE)
}
\arguments{
\item{...}{Arguments passed to \code{\link[flexdashboard:flex_dashboard]{flexdashboard::flex_dashboard()}}}

\item{display_pins}{Should the dashboard display a link to the pin(s)?
Defaults to \code{TRUE}.}

\item{display_api}{Should the dashboard display the OpenAPI documentation
for the deployed model? Defaults to \code{TRUE}.}
}
\description{
R Markdown format for model monitoring dashboards
}
\details{
The \code{modelops_dashboard()} function is a specialized type of
\pkg{flexdashboard}. See the flexdashboard website for additional
documentation:
\href{http://rmarkdown.rstudio.com/flexdashboard/}{http://rmarkdown.rstudio.com/flexdashboard/}
}
