% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelops-endpoint.R
\name{modelops_endpoint}
\alias{modelops_endpoint}
\title{Create a model API endpoint object for prediction}
\usage{
modelops_endpoint(url, ...)
}
\arguments{
\item{url}{An API endpoint URL}

\item{...}{Other arguments, not currently used}
}
\value{
A new \code{modelops_endpoint} object
}
\description{
Create a model API endpoint object for prediction
}
\examples{

\dontrun{
modelops_endpoint("https://colorado.rstudio.com/rsc/traffic-crashes/predict")
}

}
