% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelops-pin-router.R
\name{modelops_pin_router}
\alias{modelops_pin_router}
\title{Create a Plumber API to predict with a deployable \code{modelops()} object}
\usage{
modelops_pin_router(
  pr,
  modelops,
  path = "/predict",
  debug = interactive(),
  ...
)
}
\arguments{
\item{pr}{A Plumber router, such as from \code{\link[plumber:pr]{plumber::pr()}}.}

\item{modelops}{A deployable model object created with \code{\link[=modelops]{modelops()}}}

\item{path}{The endpoint path}

\item{debug}{\code{TRUE} provides more insight into your API errors.}

\item{...}{Other arguments passed to \code{predict()}, such as prediction \code{type}}
}
\description{
Use \code{modelops_pin_router()} to pin a \code{\link[=modelops]{modelops()}} trained model object to a
board of models \strong{and} set up a Plumber router with a POST endpoint for
predictions from the trained model.
}
\details{
Use \code{\link[pins:pin_read]{pins::pin_read()}} to retrieve the stored, versioned model
from the pins board by name, if needed.
}
\examples{
library(pins)
model_board <- board_temp()

cars_lm <- lm(mpg ~ ., data = mtcars)
m <- modelops(cars_lm, "cars_linear", model_board)

library(plumber)
pr() \%>\% modelops_pin_router(m)
## next, pipe to `pr_run()`

}
