% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelops-pin-write.R
\name{modelops_pin_write}
\alias{modelops_pin_write}
\title{Write a trained model to a board of models}
\usage{
modelops_pin_write(modelops)
}
\arguments{
\item{modelops}{A deployable model object created with \code{\link[=modelops]{modelops()}}}
}
\description{
Use \code{modelops_pin_write()} to pin a trained model to a board of models,
along with an input prototype for new data and other model metadata.
}
\details{
This function creates a pin on the specified \code{board} containing
the model object itself and other elements needed for prediction, such as
the model's input data prototype or which packages are needed at prediction
time. Use \code{\link[pins:pin_read]{pins::pin_read()}} to retrieve the stored, versioned model by
name from the board.
}
\examples{
library(pins)
model_board <- board_temp()

cars_lm <- lm(mpg ~ ., data = mtcars)
m <- modelops(cars_lm, "cars_linear", model_board)
modelops_pin_write(m)
model_board

}
