% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelops-pr-deploy.R
\name{modelops_pr_deploy}
\alias{modelops_pr_deploy}
\title{Add a POST endpoint to a Plumber router using a pinned model workflow object}
\usage{
modelops_pr_deploy(
  pr,
  modelops,
  type = NULL,
  path = "/predict",
  debug = interactive()
)
}
\arguments{
\item{pr}{A Plumber router, such as from \code{\link[plumber:pr]{plumber::pr()}}.}

\item{modelops}{A deployable model object created with \code{\link[=modelops]{modelops()}}.}

\item{type}{A single character or \code{NULL} describing the type of prediction.
The specific pinned model must support the \code{type} requested. Some examples
of \code{type} for \code{\link[workflows:workflow]{workflows::workflow()}} include "class", "prob", and "numeric".
When \code{NULL}, a default prediction type will be chosen based on the model
characteristics.}

\item{path}{The endpoint path}

\item{debug}{\code{TRUE} provides more insight into your API errors.}
}
\description{
Models that have been pinned to a board with \code{\link[=modelops_pin_write]{modelops_pin_write()}} can be added to a
Plumber router as a POST handler. The argument \code{type} specifies what kind of
predictions the handler will return.
}
\examples{
library(pins)
model_board <- board_temp()

cars_lm <- lm(mpg ~ ., data = mtcars)
m <- modelops(cars_lm, "cars_linear", model_board)

library(plumber)
pr() \%>\% modelops_pr_deploy(m)
## next, pipe to `pr_run()`

}
