% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelops-pr-predict.R
\name{modelops_pr_predict}
\alias{modelops_pr_predict}
\title{Create a Plumber API to predict with a deployable \code{modelops()} object}
\usage{
modelops_pr_predict(
  pr,
  modelops,
  path = "/predict",
  debug = interactive(),
  ...
)
}
\arguments{
\item{pr}{A Plumber router, such as from \code{\link[plumber:pr]{plumber::pr()}}.}

\item{modelops}{A deployable model object created with \code{\link[=modelops]{modelops()}}}

\item{path}{The endpoint path}

\item{debug}{\code{TRUE} provides more insight into your API errors.}

\item{...}{Other arguments passed to \code{predict()}, such as prediction \code{type}}
}
\description{
Use \code{modelops_pr_predict()} to add a POST endpoint for predictions from a
trained, pinned \code{\link[=modelops]{modelops()}} object to a Plumber router.
}
\details{
First store and version your \code{\link[=modelops]{modelops()}} object with
\code{\link[=modelops_pin_write]{modelops_pin_write()}}, and then create an API endpoint with
\code{modelops_pr_predict()}.

Setting \code{debug = TRUE} may expose any sensitive data from your model in
API errors.
}
\examples{
library(pins)
model_board <- board_temp()

cars_lm <- lm(mpg ~ ., data = mtcars)
m <- modelops(cars_lm, "cars_linear", model_board)
modelops_pin_write(m)

library(plumber)
pr() \%>\% modelops_pr_predict(m)
## next, pipe to `pr_run()`

}
