% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_model.R
\name{pin_model}
\alias{pin_model}
\title{Pin a trained model to a board of models}
\usage{
pin_model(
  board,
  model,
  model_id,
  type = "rds",
  desc = NULL,
  metadata = NULL,
  ptype = TRUE,
  versioned = TRUE
)
}
\arguments{
\item{board}{A pin board, created by \code{\link[pins:board_folder]{pins::board_folder()}},
\code{\link[pins:board_rsconnect]{pins::board_rsconnect()}}, or other \code{board_} function from the pins package.}

\item{model}{A trained model, such as a modeling \code{\link[workflows:workflow]{workflows::workflow()}}.}

\item{model_id}{Model ID or name.}

\item{type}{Defaults to \code{"rds"}, which is appropriate for most models. (See
\code{\link[pins:pin_read]{pins::pin_write()}} for other options.)}

\item{desc}{A text description of the model; most important for shared
boards so that others can understand what the model is. If omitted,
the package will generate a brief description of the contents.}

\item{metadata}{A list containing additional metadata to store with the pin.
When retrieving the pin, this will be stored in the \code{user} key, to
avoid potential clashes with the metadata that pins itself uses.}

\item{ptype}{Should an input data prototype be stored with the model?
Defaults to \code{TRUE}.}

\item{versioned}{Should the model object be versioned? Defaults to \code{TRUE}.}
}
\description{
Use \code{pin_model()} to pin a trained model to a board of models, along with an
input prototype for new data and other model metadata.
}
\details{
This function creates a pin on the specified \code{board} containing
the model object itself and other elements needed for prediction, such as the
model's input data prototype or which packages are needed at prediction time.
}
\examples{
library(pins)
model_board <- board_temp()

cars_lm <- lm(mpg ~ ., data = mtcars)

model_board \%>\% pin_model(cars_lm, "cars")
model_board

}
