% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pr_model}
\alias{pr_model}
\title{Add a POST endpoint to a Plumber router using a pinned model workflow object}
\usage{
pr_model(pr, board, model_id, type = NULL, path = "/predict", ...)
}
\arguments{
\item{pr}{A Plumber router, such as from \code{\link[plumber:pr]{plumber::pr()}}.}

\item{board}{A board containing models pinned via \code{\link{pin_model}}.}

\item{model_id}{Model ID or name.}

\item{type}{A single character or \code{NULL} describing the type of prediction.
The specific pinned model must support the \code{type} requested. Some examples
of \code{type} for \code{\link[workflows:workflow]{workflows::workflow()}} include "class", "prob", and "numeric".
When \code{NULL}, a default prediction type will be chosen based on the model
characteristics.}

\item{path}{The endpoint path}

\item{...}{Other arguments passed to \code{\link[plumber:pr_handle]{plumber::pr_post()}}.}
}
\description{
Models that have been pinned to a board with \code{\link{pin_model}} can be added to a
Plumber router as a POST handler. The argument \code{type} specifies what kind of
predictions the handler will return.
}
