% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_api}
\alias{predict_api}
\title{Post new data to a deployed model API endpoint and return predictions}
\usage{
predict_api(endpoint, new_data)
}
\arguments{
\item{endpoint}{URL of a prediction endpoint, such as one created by
\code{\link[=pr_model]{pr_model()}}.}

\item{new_data}{New data for making predictions, such as a data frame.}
}
\value{
A tibble with as many rows as in \code{new_data} and model predictions.
}
\description{
Post new data to a deployed model API endpoint and return predictions
}
\examples{

\dontrun{
predict_api("http://127.0.0.1:4558/predict", mtcars[,2:11])
}


}
