% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{predict_glm}
\alias{predict_glm}
\title{Title}
\usage{
predict_glm(
  object,
  newdata = NULL,
  type = c("link", "response", "terms"),
  se.fit = FALSE,
  dispersion = NULL,
  terms = NULL,
  na.action = na.pass,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
predict glm function that gives se.fit even when data is missing
this was needed because dispersion fails to be calculated without the
original training data. dispersion is basically always = 1 though so
it's not very necessary
}
