% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vetiver-package.R
\docType{package}
\name{vetiver-package}
\alias{vetiver}
\alias{vetiver-package}
\title{vetiver: Version, Share, Deploy, and Monitor Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The goal of 'vetiver' is to provide fluent tooling to version, share, deploy, and monitor a trained model. Functions handle both recording and checking the model's input data prototype, and predicting from a remote API endpoint. The 'vetiver' package is extensible, with generics that can support many kinds of models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://vetiver.posit.co}
  \item \url{https://rstudio.github.io/vetiver-r/}
  \item \url{https://github.com/rstudio/vetiver-r/}
  \item Report bugs at \url{https://github.com/rstudio/vetiver-r/issues}
}

}
\author{
\strong{Maintainer}: Julia Silge \email{julia.silge@posit.co} (\href{https://orcid.org/0000-0002-3671-836X}{ORCID})

Authors:
\itemize{
  \item Julia Silge \email{julia.silge@posit.co} (\href{https://orcid.org/0000-0002-3671-836X}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
