% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker.R
\name{vetiver_deploy_sagemaker}
\alias{vetiver_deploy_sagemaker}
\title{Deploy a vetiver model API to RStudio Connect}
\usage{
vetiver_deploy_sagemaker(
  board,
  name,
  version = NULL,
  predict_args = list(),
  repo_name = glue("vetiver-sagemaker-{name}"),
  compute_type = c("BUILD_GENERAL1_SMALL", "BUILD_GENERAL1_MEDIUM",
    "BUILD_GENERAL1_LARGE", "BUILD_GENERAL1_2XLARGE"),
  bucket = NULL,
  instance_type = NULL,
  ...
)
}
\arguments{
\item{board}{A pin board, created by \code{\link[pins:board_folder]{board_folder()}}, \code{\link[pins:board_connect]{board_connect()}},
\code{\link[pins:board_url]{board_url()}} or another \code{board_} function.}

\item{name}{Pin name.}

\item{version}{Retrieve a specific version of a pin. Use \code{\link[pins:pin_versions]{pin_versions()}} to
find out which versions are available and when they were created.}

\item{predict_args}{A list of optional arguments passed to \code{\link[=vetiver_api]{vetiver_api()}}
such as the endpoint \code{path} or prediction \code{type}.}

\item{repo_name}{The \verb{AWS ECR} repository name to store the model.}

\item{compute_type}{The \verb{AWS CodeBuild} instance to build docker for the model.}

\item{bucket}{(character): The S3 bucket to use for sending data to \code{CodeBuild} (if None,
use the \verb{SageMaker SDK} default bucket).}

\item{instance_type}{The \verb{Amazon SageMaker} instance to host the model.}

\item{...}{Other arguments passed to \code{\link[=vetiver_sm_endpoint]{vetiver_sm_endpoint()}} such as
\code{accelerator_type} or \code{data_capture_config}.}
}
\value{
The deployment success (\code{TRUE} or \code{FALSE}), invisibly.
}
\description{
Use \code{vetiver_deploy_sagemaker()} to deploy a \code{\link[=vetiver_model]{vetiver_model()}} that has been
versioned and stored via \code{\link[=vetiver_pin_write]{vetiver_pin_write()}} as a Plumber API on
\verb{Amazon SageMaker}.
}
\details{
The two functions \code{vetiver_deploy_rsconnect()} and
\code{\link[=vetiver_create_rsconnect_bundle]{vetiver_create_rsconnect_bundle()}} are alternatives to each other, providing
different strategies for deploying a vetiver model API to RStudio Connect.
}
\seealso{
\code{\link[=vetiver_write_plumber]{vetiver_write_plumber()}}, \code{\link[=vetiver_create_rsconnect_bundle]{vetiver_create_rsconnect_bundle()}}
}
