% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{vetiver_endpoint}
\alias{vetiver_endpoint}
\title{Create a model API endpoint object for prediction}
\usage{
vetiver_endpoint(url, ...)
}
\arguments{
\item{url}{An API endpoint URL}

\item{...}{Other arguments, not currently used}
}
\value{
A new \code{vetiver_endpoint} object
}
\description{
Create a model API endpoint object for prediction
}
\examples{

\dontrun{
vetiver_endpoint("https://colorado.rstudio.com/rsc/traffic-crashes/predict")
}

}
