% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{vetiver_grid.default}
\alias{vetiver_grid.default}
\title{Title}
\usage{
\method{vetiver_grid}{default}(model, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
# lm
m <- lm(mpg ~ ., data=mtcars)
m <- lm(mpg ~ cyl*disp + hp, data=mtcars)
m <- lm(mpg ~ cyl + I(cyl*2) + disp, data=mtcars)
}
