% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vetiver-model.R
\name{vetiver_meta}
\alias{vetiver_meta}
\title{Metadata constructor for \code{vetiver_model()} object}
\usage{
vetiver_meta(user = list(), version = NULL, url = NULL, required_pkgs = NULL)
}
\arguments{
\item{user}{Metadata supplied by the user}

\item{version}{Version of the pin}

\item{url}{URL for the pin, if any}

\item{required_pkgs}{Character string of R packages required for prediction}
}
\value{
A list.
}
\description{
The metadata stored in a \code{\link[=vetiver_model]{vetiver_model()}} object has four elements:
}
\details{
\itemize{
\item \verb{$user}, the metadata supplied by the user
\item \verb{$version}, the version of the pin (which can be \code{NULL} before pinning)
\item \verb{$url}, the URL where the pin is located, if any
\item \verb{$required_pkgs}, a character string of R packages required for prediction
}
}
\examples{
vetiver_meta()

}
