% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr-predict.R
\name{vetiver_pr_predict}
\alias{vetiver_pr_predict}
\title{Create a Plumber API to predict with a deployable \code{vetiver()} object}
\usage{
vetiver_pr_predict(pr, vetiver, path = "/predict", debug = interactive(), ...)
}
\arguments{
\item{pr}{A Plumber router, such as from \code{\link[plumber:pr]{plumber::pr()}}.}

\item{vetiver}{A deployable model object created with \code{\link[=vetiver]{vetiver()}}}

\item{path}{The endpoint path}

\item{debug}{\code{TRUE} provides more insight into your API errors.}

\item{...}{Other arguments passed to \code{predict()}, such as prediction \code{type}}
}
\description{
Use \code{vetiver_pr_predict()} to add a POST endpoint for predictions from a
trained, pinned \code{\link[=vetiver]{vetiver()}} object to a Plumber router.
}
\details{
First store and version your \code{\link[=vetiver]{vetiver()}} object with
\code{\link[=vetiver_pin_write]{vetiver_pin_write()}}, and then create an API endpoint with
\code{vetiver_pr_predict()}.

Setting \code{debug = TRUE} may expose any sensitive data from your model in
API errors.
}
\examples{
library(pins)
model_board <- board_temp()

cars_lm <- lm(mpg ~ ., data = mtcars)
m <- vetiver(cars_lm, "cars_linear", model_board)
vetiver_pin_write(m)

library(plumber)
pr() \%>\% vetiver_pr_predict(m)
## next, pipe to `pr_run()`

}
