% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme.R
\name{vetiver_ptype.lme}
\alias{vetiver_ptype.lme}
\title{this is mainly need for keeping track of all the levels for factor variables
which itself is mainly needed for generating confidence intervals}
\usage{
\method{vetiver_ptype}{lme}(model, ...)
}
\description{
this is mainly need for keeping track of all the levels for factor variables
which itself is mainly needed for generating confidence intervals
}
