% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras.R, R/luz.R, R/write-plumber.R
\name{vetiver_python_requirements.keras.engine.training.Model}
\alias{vetiver_python_requirements.keras.engine.training.Model}
\alias{vetiver_renviron_requirements.luz_module_fitted}
\alias{vetiver_python_requirements}
\alias{vetiver_python_requirements.default}
\alias{vetiver_renviron_requirements}
\alias{vetiver_renviron_requirements.default}
\title{Use extra files required for deployment}
\usage{
\method{vetiver_python_requirements}{keras.engine.training.Model}(model)

\method{vetiver_renviron_requirements}{luz_module_fitted}(model)

vetiver_python_requirements(model)

\method{vetiver_python_requirements}{default}(model)

vetiver_renviron_requirements(model)

\method{vetiver_renviron_requirements}{default}(model)
}
\arguments{
\item{model}{A trained model, such as an \code{lm()} model or a tidymodels
\code{\link[workflows:workflow]{workflows::workflow()}}.}
}
\description{
Create files required for deploying an app generated via
\code{\link[=vetiver_write_plumber]{vetiver_write_plumber()}}, such as a Python \code{requirements.txt} or an
\code{.Renviron}
}
\keyword{internal}
