% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker.R
\name{vetiver_sm_delete}
\alias{vetiver_sm_delete}
\title{Delete \verb{Amazon SageMaker} endpoint and configuration.}
\usage{
vetiver_sm_delete(endpoint_name, delete_endpoint_config = TRUE)
}
\arguments{
\item{endpoint_name}{\verb{Amazon SageMaker} model endpoint to be deleted.}

\item{delete_endpoint_config}{Delete endpoint configuration (default: \code{TRUE}).}
}
\value{
\code{TRUE} invisible
}
\description{
Delete \verb{Amazon SageMaker} endpoint and configuration.
}
