% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker.R
\name{vetiver_sm_endpoint}
\alias{vetiver_sm_endpoint}
\title{Create a \verb{Amazon SageMaker} Model endpoint.}
\usage{
vetiver_sm_endpoint(
  model_name,
  endpoint_name = NULL,
  instance_type = NULL,
  initial_instance_count = 1,
  accelerator_type = NULL,
  tags = NULL,
  kms_key = NULL,
  data_capture_config = NULL,
  volume_size = NULL,
  model_data_download_timeout = NULL,
  wait = TRUE
)
}
\arguments{
\item{model_name}{The \verb{Amazon SageMaker} model name}

\item{endpoint_name}{The name to use for \verb{Amazon SageMaker} model endpoint.}

\item{instance_type}{Type of EC2 instance to use
\href{https://aws.amazon.com/sagemaker/pricing/}{Amazon SageMaker Pricing}.}

\item{initial_instance_count}{The initial number of instances to run
in the \code{Endpoint}.}

\item{accelerator_type}{Type of Elastic Inference accelerator to
attach to an endpoint for model loading and inference, for
example, 'ml.eia1.medium'.}

\item{tags}{A list of tags for labeling \verb{Amazon SageMaker} Model.}

\item{kms_key}{The ARN of the KMS key that is used to encrypt the
data on the storage volume attached to the instance hosting the
endpoint.}

\item{data_capture_config}{A list configuration to control how \verb{Amazon SageMaker}
captures inference data.}

\item{volume_size}{The size, in GB, of the ML storage volume attached to individual
inference instance associated with the production variant. Currenly only Amazon EBS
gp2 storage volumes are supported.}

\item{model_data_download_timeout}{The timeout value, in seconds, to download and
extract model data from \verb{Amazon S3}.}

\item{wait}{Wait for the endpoint to be deployed.}
}
\value{
A new \code{\link[=vetiver_endpoint_sagemaker]{vetiver_endpoint_sagemaker()}} object
}
\description{
Create a \verb{Amazon SageMaker} Model endpoint.
}
