% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker.R
\name{vetiver_sm_model}
\alias{vetiver_sm_model}
\title{Create a \verb{Amazon SageMaker} Model.}
\usage{
vetiver_sm_model(
  image_uri,
  model_name,
  role,
  vpc_config = list(),
  enable_network_isolation = FALSE,
  tags = NULL
)
}
\arguments{
\item{image_uri}{The AWS ECR image uri for \verb{Amazon SageMaker} Model}

\item{model_name}{The \verb{Amazon SageMaker} model name (optional)}

\item{role}{The ARN role for the \verb{Amazon SageMaker} model (optional)}

\item{vpc_config}{A list containing the \code{VPC} configuration for \verb{Amazon SageMaker} model
\href{https://docs.aws.amazon.com/sagemaker/latest/APIReference/API_VpcConfig.html}{API VpcConfig} (optional).
\itemize{
\item \code{Subnets}: List of subnet ids
\item \code{SecurityGroupIds}: List of security group ids.
}}

\item{enable_network_isolation}{Specifies whether container will
run in network isolation mode.}

\item{tags}{A list of tags for labeling \verb{Amazon SageMaker} Model.}
}
\value{
Character vector of the model name
}
\description{
Create a \verb{Amazon SageMaker} Model.
}
