% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-docker.R
\name{vetiver_write_docker}
\alias{vetiver_write_docker}
\title{Generate files necessary to build a Docker container for a vetiver model}
\usage{
vetiver_write_docker(
  board,
  name,
  version = NULL,
  path = ".",
  lockfile = "vetiver_renv.lock",
  rspm = TRUE,
  port = 8000,
  expose = TRUE,
  additional_pkgs = required_pkgs(board),
  predict_args = list()
)
}
\arguments{
\item{board}{A pin board, created by \code{\link[pins:board_folder]{board_folder()}}, \code{\link[pins:board_rsconnect]{board_rsconnect()}},
\code{\link[pins:board_url]{board_url()}} or another \code{board_} function.}

\item{name}{Pin name.}

\item{version}{Retrieve a specific version of a pin. Use \code{\link[pins:pin_versions]{pin_versions()}} to
find out which versions are available and when they were created.}

\item{path}{A path to write the Plumber file, Dockerfile, and \code{lockfile},
capturing the model's dependencies. Defaults to the working directory.}

\item{lockfile}{The generated lockfile in \code{path}. Defaults to
\code{"vetiver_renv.lock"}.}

\item{rspm}{A logical to use the
\href{https://packagemanager.rstudio.com/}{RStudio Public Package Manager} for
\code{\link[renv:restore]{renv::restore()}} in the Docker image. Defaults to \code{TRUE}.}

\item{port}{The server port for listening: a number such as 8080 or an
expression like \code{'as.numeric(Sys.getenv("PORT"))'} when the port is injected
as an environment variable.}

\item{expose}{Add \code{EXPOSE} to the Dockerfile? This is helpful for using
Docker Desktop but does not work with an expression for \code{port}.}

\item{additional_pkgs}{A character vector of additional package names to add
to the Docker image. For example, some boards like \code{\link[pins:board_s3]{pins::board_s3()}} require
additional software; you can use \code{required_pkgs(board)} here.}

\item{predict_args}{A list of optional arguments passed to \code{\link[=vetiver_api]{vetiver_api()}}
such as the endpoint \code{path} or prediction \code{type}.}
}
\value{
This function is called mainly for its side effects. It returns
the \code{path} invisibly.
}
\description{
Deploying a vetiver model via Docker requires several files. Use this
function to create these needed files in the directory located at \code{path}.
}
\details{
This function uses \code{\link[=vetiver_write_plumber]{vetiver_write_plumber()}} to create a Plumber file
appropriate for Docker, along with a Dockerfile and renv \code{lockfile}.
}
\examples{
\dontshow{if (interactive() || identical(Sys.getenv("IN_PKGDOWN"), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

library(pins)
tmp_plumber <- tempfile()
b <- board_temp(versioned = TRUE)
cars_lm <- lm(mpg ~ ., data = mtcars)
v <- vetiver_model(cars_lm, "cars_linear")
vetiver_pin_write(b, v)
vetiver_write_plumber(b, "cars_linear", file = tmp_plumber)

## default port
vetiver_write_docker(b, "cars_linear", tempdir())
## pass args for predicting:
vetiver_write_docker(b, "cars_linear", tempdir(),
                     predict_args = list(debug = TRUE))
## port from env variable
vetiver_write_docker(b, "cars_linear", tempdir(),
                     port = 'as.numeric(Sys.getenv("PORT"))',
                     expose = FALSE)
\dontshow{\}) # examplesIf}
}
