% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-api-spec.R
\name{api_spec}
\alias{api_spec}
\title{Update the OpenAPI specification from model metadata}
\usage{
api_spec(spec, vetiver_model, path)
}
\arguments{
\item{spec}{An OpenAPI Specification formatted list object}

\item{vetiver_model}{A deployable \code{\link[=vetiver_model]{vetiver_model()}} object}

\item{path}{The endpoint path}
}
\value{
The new OpenAPI Specification object updated
}
\description{
Update the OpenAPI specification from model metadata
}
