% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-api-spec.R
\name{api_spec}
\alias{api_spec}
\title{Update the OpenAPI specification from model metadata}
\usage{
api_spec(spec, vetiver_model, path)
}
\arguments{
\item{spec}{An OpenAPI Specification formatted list object}

\item{vetiver_model}{A deployable \code{\link[=vetiver_model]{vetiver_model()}} object}

\item{path}{The endpoint path}
}
\value{
The updated OpenAPI Specification object
}
\description{
Update the OpenAPI specification from model metadata
}
\examples{
library(pins)
library(plumber)
model_board <- board_temp()
cars_lm <- lm(mpg ~ ., data = mtcars)
v <- vetiver_model(cars_lm, "cars_linear", model_board)

modify_spec <- function(spec) api_spec(spec, v, "/predict")
pr() \%>\% pr_set_api_spec(api = modify_spec)

}
