% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_api_spec.R
\name{api_spec}
\alias{api_spec}
\title{Update the OpenAPI specification from model metadata}
\usage{
api_spec(spec, modelops, path)
}
\arguments{
\item{spec}{An OpenAPI Specification formatted list object}

\item{modelops}{A deployable model object created with \code{\link[=modelops]{modelops()}}.}

\item{path}{The endpoint path}
}
\value{
The new OpenAPI Specification object updated
}
\description{
Update the OpenAPI specification from model metadata
}
