% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{handle_model}
\alias{handle_model}
\alias{handle_model.default}
\alias{handle_model.lm}
\title{Wrapper function for creating model handler function}
\usage{
handle_model(x, ...)

\method{handle_model}{default}(x, ...)

\method{handle_model}{lm}(x, ...)
}
\arguments{
\item{x}{A trained model}

\item{...}{Other arguments passed from \code{\link[=pr_model]{pr_model()}}}
}
\description{
Wrapper function for creating model handler function
}
