% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelops-pin-write.R
\name{modelops_pin_write}
\alias{modelops_pin_write}
\title{Write a trained model to a board of models}
\usage{
modelops_pin_write(modelops)
}
\arguments{
\item{modelops}{A deployable model object created with \code{\link[=modelops]{modelops()}}}
}
\description{
Use \code{modelops_pin_write()} to pin a \code{\link[=modelops]{modelops()}} trained model object (plus
its input data prototype and other metadata) to a board of models.
}
\details{
This function is used by \code{\link[=modelops_pin_router]{modelops_pin_router()}} to create a pin
on the object's \code{board} containing the model object itself and other elements
needed for prediction, such as the model's input data prototype or which
packages are needed at prediction time. This function may also be used alone,
if you would like to \emph{store} and \emph{version} a model but not set up an API
endpoint.

Use \code{\link[pins:pin_read]{pins::pin_read()}} to retrieve the stored, versioned model by name from
the board.
}
\examples{
library(pins)
model_board <- board_temp()

cars_lm <- lm(mpg ~ ., data = mtcars)
m <- modelops(cars_lm, "cars_linear", model_board)
modelops_pin_write(m)
model_board

}
