% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vetiver-package.R
\docType{package}
\name{vetiver-package}
\alias{vetiver}
\alias{vetiver-package}
\title{vetiver: Version, Pin, Deploy, and Monitor Models with 'pins' and 'plumber'}
\description{
A trained model can be shared and versioned via 'pins', deployed as an API via 'plumber', and monitored via a dashboard created with 'rmarkdown'. Functions handle recording and checking the model's input data prototype, predicting from a remote API endpoint, and monitoring model performance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tidymodels.github.io/vetiver}
  \item \url{https://github.com/tidymodels/vetiver}
  \item Report bugs at \url{https://github.com/tidymodels/vetiver/issues}
}

}
\author{
\strong{Maintainer}: Julia Silge \email{julia.silge@rstudio.com} (\href{https://orcid.org/0000-0002-3671-836X}{ORCID})

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
