% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-ptype.R, R/lm.R, R/xgboost.R
\name{vetiver_create_ptype}
\alias{vetiver_create_ptype}
\alias{vetiver_slice_zero}
\alias{vetiver_slice_zero.default}
\alias{vetiver_slice_zero.lm}
\alias{vetiver_slice_zero.xgb.Booster}
\title{Create a vetiver input data prototype}
\usage{
vetiver_create_ptype(model, save_ptype, ...)

vetiver_slice_zero(model, ...)

\method{vetiver_slice_zero}{default}(model, ...)

\method{vetiver_slice_zero}{lm}(model, ...)

\method{vetiver_slice_zero}{xgb.Booster}(model, ...)
}
\arguments{
\item{model}{A trained model, such as an \code{lm()} model or a tidymodels
\code{\link[workflows:workflow]{workflows::workflow()}}.}

\item{save_ptype}{Should an input data prototype be stored with the model?
The options are \code{TRUE} (the default, which stores a zero-row slice of the
training data), \code{FALSE} (no input data prototype for checking), or a
dataframe to be used for both checking at prediction time \emph{and} examples in
API visual documentation.}

\item{...}{Other method-specific arguments passed to \code{\link[=vetiver_slice_zero]{vetiver_slice_zero()}}
to compute an input data prototype.}
}
\value{
Either a zero-row dataframe, \code{NULL}, or the dataframe passed to
\code{save_ptype}.
}
\description{
Optionally find and return an input data prototype for a model.
}
\details{
These are developer-facing functions, useful for supporting new model types.
A \code{\link[=vetiver_model]{vetiver_model()}} object optionally stores an input data prototype for
checking at prediction time.
\itemize{
\item The default for \code{save_ptype}, \code{TRUE}, finds a zero-row slice of the
training data via \code{\link[=vetiver_slice_zero]{vetiver_slice_zero()}}.
\item \code{save_ptype = FALSE} opts out of storing any input data prototype.
\item You may pass your own data to \code{save_ptype}, but be sure to check that it
has the same structure as your training data, perhaps with
\code{\link[hardhat:scream]{hardhat::scream()}}.
}
}
\examples{

cars_lm <- lm(mpg ~ cyl + disp, data = mtcars)

vetiver_create_ptype(cars_lm, TRUE)

## calls the right method for `model` via:
vetiver_slice_zero(cars_lm)

## can also turn off `ptype`
vetiver_create_ptype(cars_lm, FALSE)

}
