% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{vetiver_dashboard}
\alias{vetiver_dashboard}
\alias{get_vetiver_dashboard_pins}
\alias{pin_example_kc_housing_model}
\title{R Markdown format for model monitoring dashboards}
\usage{
vetiver_dashboard(pins, display_pins = TRUE, ...)

get_vetiver_dashboard_pins()

pin_example_kc_housing_model(board = pins::board_temp(), name = "seattle_rf")
}
\arguments{
\item{pins}{A list containing \code{board}, \code{name}, and \code{version}, as in
\code{\link[pins:pin_read]{pins::pin_read()}}}

\item{display_pins}{Should the dashboard display a link to the pin(s)?
Defaults to \code{TRUE}.}

\item{...}{Arguments passed to \code{\link[flexdashboard:flex_dashboard]{flexdashboard::flex_dashboard()}}}
}
\description{
R Markdown format for model monitoring dashboards
}
\details{
The \code{vetiver_dashboard()} function is a specialized type of
\pkg{flexdashboard}. See the flexdashboard website for additional
documentation:
\href{http://rmarkdown.rstudio.com/flexdashboard/}{http://rmarkdown.rstudio.com/flexdashboard/}
}
