% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{vetiver_dashboard}
\alias{vetiver_dashboard}
\title{R Markdown format for model monitoring dashboards}
\usage{
vetiver_dashboard(board, name, version = NULL, ..., display_pins = TRUE)
}
\arguments{
\item{board}{A pin board, created by \code{\link[pins:board_folder]{board_folder()}}, \code{\link[pins:board_rsconnect]{board_rsconnect()}},
\code{\link[pins:board_url]{board_url()}} or another \code{board_} function.}

\item{name}{Pin name.}

\item{version}{Retrieve a specific version of a pin. Use \code{\link[pins:pin_versions]{pin_versions()}} to
find out which versions are available and when they were created.}

\item{...}{Arguments passed to \code{\link[flexdashboard:flex_dashboard]{flexdashboard::flex_dashboard()}}}

\item{display_pins}{Should the dashboard display a link to the pin(s)?
Defaults to \code{TRUE}.}
}
\description{
R Markdown format for model monitoring dashboards
}
\details{
The \code{vetiver_dashboard()} function is a specialized type of
\pkg{flexdashboard}. See the flexdashboard website for additional
documentation:
\href{http://rmarkdown.rstudio.com/flexdashboard/}{http://rmarkdown.rstudio.com/flexdashboard/}
}
