% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{vetiver_endpoint}
\alias{vetiver_endpoint}
\title{Create a model API endpoint object for prediction}
\usage{
vetiver_endpoint(url)
}
\arguments{
\item{url}{An API endpoint URL}
}
\value{
A new \code{vetiver_endpoint} object
}
\description{
Create a model API endpoint object for prediction
}
\examples{
vetiver_endpoint("https://colorado.rstudio.com/rsc/biv_svm_api/predict")

}
