% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker.R
\name{vetiver_endpoint_sagemaker}
\alias{vetiver_endpoint_sagemaker}
\title{Create a SageMaker model API endpoint object for prediction}
\usage{
vetiver_endpoint_sagemaker(model_endpoint)
}
\arguments{
\item{model_endpoint}{The name of the Amazon SageMaker model endpoint.}
}
\value{
A new \code{vetiver_endpoint_sagemaker} object
}
\description{
This function creates a model API endpoint for prediction from a Sagemaker Model.
No HTTP calls are made until you actually
\code{\link[=predict.vetiver_endpoint_sagemaker]{predict()}} with your endpoint.
}
\examples{
\dontshow{if (rlang::is_installed("smdocker")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
vetiver_endpoint_sagemaker("vetiver-sagemaker-demo-model")
\dontshow{\}) # examplesIf}
}
