% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr-post.R
\name{vetiver_pr_predict}
\alias{vetiver_pr_predict}
\title{Create a Plumber API to predict with a deployable \code{vetiver_model()} object}
\usage{
vetiver_pr_predict(
  pr,
  vetiver_model,
  path = "/predict",
  debug = is_interactive(),
  ...
)
}
\arguments{
\item{pr}{A Plumber API. Note: The supplied Plumber API object will also be updated in place as well as returned by the function.}

\item{path}{The endpoint path}

\item{...}{Additional arguments for \code{PlumberEndpoint}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated to use \link{vetiver_pr_post} directly instead.
}
\details{
Use \code{vetiver_pr_predict()} to add a POST endpoint for predictions from a
trained \code{\link[=vetiver_model]{vetiver_model()}} to a Plumber router.
}
\keyword{internal}
