% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker.R
\name{vetiver_sm_delete}
\alias{vetiver_sm_delete}
\title{Delete \verb{Amazon SageMaker} endpoint and configuration.}
\usage{
vetiver_sm_delete(object, delete_endpoint_config = TRUE)
}
\arguments{
\item{object}{A model API endpoint object created with \code{\link[=vetiver_endpoint_sagemaker]{vetiver_endpoint_sagemaker()}}.}

\item{delete_endpoint_config}{Delete endpoint configuration (default: \code{TRUE}).}
}
\value{
\code{TRUE} invisible
}
\description{
Delete \verb{Amazon SageMaker} endpoint and configuration.
}
